requirements.txt
selenium==4.22.0
pytest==8.2.2
assertpy==1.1
dnspython==2.6.1
pytest-html==4.1.1
allure-pytest==2.13.5
python-dotenv==1.0.1
pytest-xdist==3.6.1
beautifulsoup4==4.12.3